# scratch/swagger.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "PufferPanel API interface for both the panel and daemon. Endpoints starting with /daemon or /proxy are for nodes.",
        "title": "PufferPanel API",
        "contact": {
            "name": "PufferPanel",
            "url": "https://pufferpanel.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "",
    "basePath": "",
    "paths": {
...
        "/api/users": {
            "get": {
                "description": "Gets users, and allowing for filtering of users. * is a wildcard that can be used for text inputs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get users",
                "parameters": [
                    {
                        "description": "Filters to search on",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "New user information",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserView"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/daemon/server/{id}/install": {
            "post": {
                "description": "installs the given server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Installs server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server Identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Wait for the operation to complete",
                        "name": "wait",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Install has been queued",
                        "schema": {
                            "$ref": "#/definitions/response.Empty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Empty"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Empty"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.UserSearch": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "pageLimit": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserSearchResponse": {
            "type": "object",
            "properties": {
                "paging": {
                    "$ref": "#/definitions/response.Paging"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserView"
                    }
                }
            }
        },
...
        "models.UserView": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "newPassword": {
                    "type": "string"
                },
                "password": {
                    "description": "ONLY SHOW WHEN COPYING",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "pufferpanel.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "response.Empty": {
            "type": "object"
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/pufferpanel.Error"
                }
            }
        },
        "response.Paging": {
            "type": "object",
            "properties": {
                "maxSize": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}